change to float

+++ javascript/Ex03.js
@@ -1,5 +1,5 @@
-var n1 = Number.parseInt(window.prompt('Digite Um número: '));
-var n2 = Number.parseInt(window.prompt('Digite Outro número: '));
+var n1 = Number.parseFloat(window.prompt('Digite Um número: '));
+var n2 = Number.parseFloat(window.prompt('Digite Outro número: '));
 var s = n1 + n2;
 window.alert('A soma dos valores é ' + s); 
 //number + number = adição
